+++ src/main/java/com/alkemy/ong/data/entities/ActivityEntity.java
@@ -1,5 +1,4 @@
-package com.alkemy.ong.data.entity;
-
+package com.alkemy.ong.data.entities;
 
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.SQLDelete;
@@ -10,34 +9,22 @@ import javax.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "users")
-@SQLDelete(sql = "UPDATE users SET is_deleted = true WHERE id=?")
+@Table(name = "activities")
+@SQLDelete(sql = "UPDATE activity SET deleted = true WHERE id=?")
 @Where(clause = "is_deleted=false")
-public class UserEntity {
+public class ActivityEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
-  @Column(name = "first_name", nullable = false)
-  private String firstName;
-
-  @Column(name = "last_name", nullable = false)
-  private String lastName;
-
-  @Column(unique = true, nullable = false)
-  private String email;
+  @Column(nullable = false)
+  private String name;
 
   @Column(nullable = false)
-  private String password;
+  private String content;
 
   @Column(nullable = false)
-  private String photo;
-
-  @ManyToOne
-  private RoleEntity role;
-
-  @Column(name = "is_deleted", nullable = false)
-  private boolean isDeleted = Boolean.FALSE;
+  private String image;
 
   @Column(name = "created_at", updatable = false, nullable = false)
   @CreationTimestamp
@@ -46,4 +33,7 @@ public class UserEntity {
   @Column(name = "updated_at", nullable = false)
   @UpdateTimestamp
   private LocalDateTime updatedAt;
+
+  @Column(name = "is_deleted", nullable = false)
+  private boolean isDeleted = Boolean.FALSE;
 }
