+++ src/main/java/com/alkemy/ong/data/entities/UserEntity.java
@@ -1,4 +1,5 @@
-package com.alkemy.ong.data.entity;
+package com.alkemy.ong.data.entities;
+
 
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.SQLDelete;
@@ -9,22 +10,34 @@ import javax.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "activities")
-@SQLDelete(sql = "UPDATE activity SET deleted = true WHERE id=?")
+@Table(name = "users")
+@SQLDelete(sql = "UPDATE users SET is_deleted = true WHERE id=?")
 @Where(clause = "is_deleted=false")
-public class ActivityEntity {
+public class UserEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
-  @Column(nullable = false)
-  private String name;
+  @Column(name = "first_name", nullable = false)
+  private String firstName;
+
+  @Column(name = "last_name", nullable = false)
+  private String lastName;
+
+  @Column(unique = true, nullable = false)
+  private String email;
 
   @Column(nullable = false)
-  private String content;
+  private String password;
 
   @Column(nullable = false)
-  private String image;
+  private String photo;
+
+  @ManyToOne
+  private RoleEntity role;
+
+  @Column(name = "is_deleted", nullable = false)
+  private boolean isDeleted = Boolean.FALSE;
 
   @Column(name = "created_at", updatable = false, nullable = false)
   @CreationTimestamp
@@ -33,7 +46,4 @@ public class ActivityEntity {
   @Column(name = "updated_at", nullable = false)
   @UpdateTimestamp
   private LocalDateTime updatedAt;
-
-  @Column(name = "is_deleted", nullable = false)
-  private boolean isDeleted = Boolean.FALSE;
 }
