+++ src/main/java/com/alkemy/ong/data/gateway/DefaultGateway.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.data.gateway;
 
-import com.alkemy.ong.data.entity.NewsEntity;
+import com.alkemy.ong.data.entities.NewsEntity;
 import com.alkemy.ong.data.exception.ParamNotFoundException;
 import com.alkemy.ong.data.mapper.NewsModelMapper;
 import com.alkemy.ong.data.repository.NewsRepository;
