+++ src/main/java/com/alkemy/ong/data/repository/NewsRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.data.repository;
 
-import com.alkemy.ong.data.entity.NewsEntity;
+import com.alkemy.ong.data.entities.NewsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
