+++ src/main/java/com/alkemy/ong/data/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.data.repository;
 
-import com.alkemy.ong.data.entity.RoleEntity;
+import com.alkemy.ong.data.entities.RoleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
