+++ src/main/java/com/alkemy/ong/data/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.data.repository;
 
-import com.alkemy.ong.data.entity.UserEntity;
+import com.alkemy.ong.data.entities.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository  extends JpaRepository<UserEntity, Long> {
